Move: auth/registerDeviceToken to notifications/registerDeviceToken
