+++ src/routes/auth.js
@@ -19,14 +19,6 @@ router.route("/app/token/refresh").get(mobileAuthHandlers.refreshAccessToken);
 router.route("/app/device").post(mobileAuthHandlers.registerDeviceTokenHandler);
 router.route("/app/device").delete(mobileAuthHandlers.removeDeviceTokenHandler);
 router.route("/app/token/generate").get(authHandlers.generateTokenHandler);
-// FCM 토큰 등록
-router.post(
-  "/registerDeviceToken",
-  authMiddleware,
-  [body("deviceToken").isString().isLength({ min: 1, max: 1024 })],
-  validator,
-  authHandlers.registerDeviceTokenHandler
-);
 
 // 환경변수 SPARCSSSO_CLIENT_ID 유무에 따라 로그인 방식이 변경됩니다.
 module.exports = sparcsssoEnv?.id ? router : authReplace;
