+++ src/routes/auth.replace.js
@@ -15,13 +15,4 @@ router.route("/sparcssso").get(authReplaceHandlers.sparcsssoHandler);
 // 로그아웃
 router.route("/logout").get(authReplaceHandlers.logoutHandler);
 
-// FCM 토큰 등록
-router.post(
-  "/registerDeviceToken",
-  authMiddleware,
-  [body("deviceToken").isString().isLength({ min: 1, max: 1024 })],
-  validator,
-  authReplaceHandlers.registerDeviceTokenHandler
-);
-
 module.exports = router;
