+++ src/routes/notifications.js
@@ -8,6 +8,16 @@ const validator = require("../middlewares/validator");
 // 라우터 접근 시 로그인 필요
 router.use(require("../middlewares/auth"));
 
+// FCM 토큰 등록
+router.post(
+  "/registerDeviceToken",
+  [body("deviceToken").isString().isLength({ min: 1, max: 1024 })],
+  validator,
+  notificationHandlers.registerDeviceTokenHandler
+);
+
+// @TODO: "/editOptions"
+
 router.get(
   "/options",
   query("deviceToken").isString(),
