+++ src/services/notifications.js
@@ -1,24 +1,62 @@
+const { userModel } = require("../modules/stores/mongo");
 const { notificationOptionModel } = require("../modules/stores/mongo");
 const logger = require("../modules/logger");
+
+const { registerDeviceToken, validateDeviceToken } = require("../modules/fcm");
+
+const registerDeviceTokenHandler = async (req, res) => {
+  try {
+    // 해당 FCM device token이 유효한지 검사합니다.
+    const { deviceToken } = req.body;
+    const isValid = await validateDeviceToken(deviceToken);
+    if (!isValid) {
+      return res
+        .status(400)
+        .send("Notifications/registerDeviceToken : deviceToken is invalid");
+    }
+
+    // 데이터베이스에 deviceToken 레코드를 추가합니다.
+    const user = await userModel.findOne({ id: req.userId }, "_id");
+    const newDeviceToken = await registerDeviceToken(user._id, deviceToken);
+
+    // 세션에 현재 사용자 기기의 deviceToken을 저장합니다.
+    req.session.deviceToken = deviceToken;
+
+    return res.status(200).json({
+      deviceToken: newDeviceToken,
+    });
+  } catch (e) {
+    logger.error(e);
+    res
+      .status(500)
+      .send("Notifications/registerDeviceToken : internal server error");
+  }
+};
 
 const optionsHandler = async (req, res) => {
   try {
     const { deviceToken } = req.query;
 
+    // 세션에 저장된 deviceToken과 요청에 들어온 deviceToken이 일치하는지 검사합니다.
+    if (req.session.deviceToken !== deviceToken) {
+      return res
+        .status(400)
+        .send("Notifications/options : deviceToken is invalid");
+    }
+
     // deviceToken에 대응되는 알림 설정을 찾아 반환합니다.
     const notificationOptions = await notificationOptionModel
       .findOne(
-        {
-          deviceToken,
-        },
+        { deviceToken },
         "-_id chatting keywords beforeDepart notice advertisement"
       )
       .lean();
     if (!notificationOptions) {
       return res
         .status(400)
-        .send("Notificaiton/getNotificationOptions: deviceToken not found");
+        .send("Notificaiton/options: deviceToken not found");
     }
+
     res.status(200).json(notificationOptions);
   } catch (err) {
     logger.error(err);
@@ -86,6 +124,7 @@ const editOptionsHandler = async (req, res) => {
 };
 
 module.exports = {
+  registerDeviceTokenHandler,
   optionsHandler,
   editOptionsHandler,
 };
